+++ jpro-routing/core/src/main/scala/one/jpro/platform/routing/Redirect.scala
@@ -1,3 +1,3 @@
 package one.jpro.platform.routing
 
-case class Redirect(to: String) extends Response
+case class Redirect(to: String) extends ResponseResult