+++ jpro-routing/core/src/main/scala/one/jpro/platform/routing/Response.scala
@@ -1,10 +1,16 @@
 package one.jpro.platform.routing
 
-trait Response
+import simplefx.experimental.FXFuture
+
+case class Response(future: FXFuture[ResponseResult]) {
+  assert(future != null, "future must not be null - but it's value can be null")
+}
 object Response {
-  def empty(): Response = null
-  def redirect(to: String): Response = Redirect(to)
-  def fromNode(node: javafx.scene.Node): Response = new Response {
-    RouteUtils.viewFromNode(node)
-  }
+  def empty(): Response = Response(FXFuture.unit(null))
+  def redirect(to: String): Response = Response(FXFuture.unit(Redirect(to)))
+
+  def view(view: View): Response = Response(FXFuture.unit(view))
+  def node(node: javafx.scene.Node): Response = Response(FXFuture.unit(View.fromNode(node)))
+  def fromFuture(future: FXFuture[Response]): Response = Response(future.flatMap(_.future))
+  def fromFutureResult(future: FXFuture[ResponseResult]): Response = Response(future)
 }