+++ jpro-routing/core/src/main/scala/one/jpro/platform/routing/Route.scala
@@ -6,34 +6,27 @@ import java.util.function.Predicate
 
 
 object Route {
-  def empty(): Route = (r) => null
+  def empty(): Route = (r) => Response.empty()
 }
 @FunctionalInterface
 trait Route {
-  def apply(r: Request): FXFuture[Response]
+  def apply(r: Request): Response
 
   def and(x: Route): Route = { request =>
     val r = apply(request)
-    if(r == null) {
-      x.apply(request)
-    } else {
-      r.flatMap{ r =>
-        if(r == null) {
-          val r = x.apply(request)
-          if(r == null) {
-            FXFuture.unit(null)
-          } else {
-            r
-          }
-        } else FXFuture.unit(r)
-      }
-    }
+    assert(r != null, "Route returned null: " + this + " for " + request)
+    Response(r.future.flatMap{ r =>
+      if(r == null) {
+        val r2 = x.apply(request)
+        r2.future
+      } else FXFuture.unit(r)
+    })
   }
   def domain(domain: String, route: Route): Route = and((r: Request) => {
     if(r.domain == domain) {
       route.apply(r)
     } else {
-      null
+      Response.empty()
     }
   })
   def path(path: String, route: Route): Route = and((r: Request) => {
@@ -41,7 +34,7 @@ trait Route {
       val r2 = r.copy(path = r.path.drop(path.length), directory = r.resolve(path))
       route.apply(r2)
     } else {
-      null
+      Response.empty()
     }
   })
   def filter(filter: Filter): Route = filter(this)
@@ -54,25 +47,28 @@ trait Route {
   }
   def filterWhenFuture(cond: Predicate[Request], filter: (Request) => FXFuture[Filter]): Route = { r =>
     if(cond.test(r)) {
-      filter(r).flatMap(filter => filter(this).apply(r))
+      Response(filter(r).flatMap(filter => filter(this).apply(r).future))
     } else {
       this.apply(r)
     }
   }
   def when(cond: Predicate[Request], _then: Route): Route = and(r => {
     val condResult = cond.test(r)
-    val r2: FXFuture[Response] = if(condResult) _then(r) else null
-    r2
+    if(condResult) _then(r) else Response.empty()
   })
   def when(cond: Predicate[Request], _then: Route, _else: Route): Route = and(r => {
     if(cond.test(r)) _then(r) else _else(r)
   })
 
   def whenFuture(cond: java.util.function.Function[Request, FXFuture[java.lang.Boolean]], _then: Route): Route = and(r => {
-    cond.apply(r).flatMap(condResult => if (condResult) _then(r) else null)
+    Response.fromFuture(
+      cond.apply(r).map(condResult => if (condResult) _then(r) else Response.empty())
+    )
   })
 
   def whenFuture(cond: java.util.function.Function[Request, FXFuture[java.lang.Boolean]], _then: Route, _else: Route): Route = and(r => {
-    cond.apply(r).flatMap(condResult => if (condResult) _then(r) else _else(r))
+    Response.fromFuture(
+      cond.apply(r).map(condResult => if (condResult) _then(r) else _else(r))
+    )
   })
 }